requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
openai>=1.52.0
aiofiles==24.1.0
python-dotenv==1.0.1
